backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Kirill Adayev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
pytest = "^8.0.0"
uvicorn = "^0.22.0"
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core=^1.7.1"]
build-backend = "poetry.core.masonry.api"
